+++ exercises/practice/armstrong-numbers/armstrong_numbers.zig
@@ -1,3 +1,3 @@
-pub fn isArmstrongNumber(num: u64) bool {
+pub fn isArmstrongNumber(num: u128) bool {
     @panic("please implement the isArmstrongNumber function");
 }
